+++ boltons/listutils.py
@@ -114,6 +114,8 @@ class BarrelList(list):
             if rel_idx < len_list:
                 break
             rel_idx -= len_list
+        else:
+            return None, None  # index >= len(self)
         if rel_idx < 0:
             return None, None
         return list_idx, rel_idx
@@ -137,9 +139,13 @@ class BarrelList(list):
             self.lists[0].insert(index, item)
             self._balance_list(0)
         else:
-            list_idx, rel_idx = self._translate_index(index)
-            if list_idx is None:
+            len_self = len(self)
+            if index >= len_self:  # clamp to append, like list.insert
+                list_idx, rel_idx = len(self.lists) - 1, len(self.lists[-1])
+            elif index < -len_self:  # clamp to prepend, like list.insert
                 list_idx, rel_idx = 0, 0
+            else:
+                list_idx, rel_idx = self._translate_index(index)
             self.lists[list_idx].insert(rel_idx, item)
             self._balance_list(list_idx)
         return
@@ -156,6 +162,8 @@ class BarrelList(list):
             return self.lists[0].pop()
         index = a and a[0]
         if index == () or index is None or index == -1:
+            while len(lists) > 1 and not lists[-1]:
+                lists.pop()  # clean up empty tail sublists first
             ret = lists[-1].pop()
             if len(lists) > 1 and not lists[-1]:
                 lists.pop()
@@ -164,27 +172,18 @@ class BarrelList(list):
             if list_idx is None:
                 raise IndexError()
             ret = lists[list_idx].pop(rel_idx)
-            self._balance_list(list_idx)
+            if len(lists) > 1 and not lists[list_idx]:
+                del lists[list_idx]  # don't accumulate empty sublists
+            else:
+                self._balance_list(list_idx)
         return ret
 
     def iter_slice(self, start, stop, step=None):
-        iterable = self  # TODO: optimization opportunities abound
-        # start_list_idx, stop_list_idx = 0, len(self.lists)
-        if start is None:
-            start = 0
-        if stop is None:
-            stop = len(self)
-        if step is not None and step < 0:
-            step = -step
-            start, stop = -start, -stop - 1
-            iterable = reversed(self)
-        if start < 0:
-            start += len(self)
-            # start_list_idx, start_rel_idx = self._translate_index(start)
-        if stop < 0:
-            stop += len(self)
-            # stop_list_idx, stop_rel_idx = self._translate_index(stop)
-        return islice(iterable, start, stop, step)
+        length = len(self)
+        start, stop, step = slice(start, stop, step).indices(length)
+        if step < 0:
+            return islice(reversed(self), length - 1 - start, length - 1 - stop, -step)
+        return islice(self, start, stop, step)
 
     def del_slice(self, start, stop, step=None):
         if step is not None and abs(step) > 1:  # punt
@@ -195,10 +194,11 @@ class BarrelList(list):
             return
         if start is None:
             start = 0
-        if stop is None:
-            stop = len(self)
         start_list_idx, start_rel_idx = self._translate_index(start)
-        stop_list_idx, stop_rel_idx = self._translate_index(stop)
+        if stop is None or stop >= len(self):
+            stop_list_idx, stop_rel_idx = len(self.lists) - 1, len(self.lists[-1])
+        else:
+            stop_list_idx, stop_rel_idx = self._translate_index(stop)
         if start_list_idx is None:
             raise IndexError()
         if stop_list_idx is None:
@@ -207,9 +207,9 @@ class BarrelList(list):
         if start_list_idx == stop_list_idx:
             del self.lists[start_list_idx][start_rel_idx:stop_rel_idx]
         elif start_list_idx < stop_list_idx:
-            del self.lists[start_list_idx + 1:stop_list_idx]
-            del self.lists[start_list_idx][start_rel_idx:]
             del self.lists[stop_list_idx][:stop_rel_idx]
+            del self.lists[start_list_idx][start_rel_idx:]
+            del self.lists[start_list_idx + 1:stop_list_idx]
         else:
             assert False, ('start list index should never translate to'
                            ' greater than stop list index')