+++ boltons/mathutils.py
@@ -161,7 +161,7 @@ class Bits:
             raise ValueError('Bits cannot represent negative values')
         if len_ is None:
             len_ = len(f'{val:b}')
-        if val > 2 ** len_:
+        if val >= 2 ** len_:
             raise ValueError(f'value {val} cannot be represented with {len_} bits')
         self.val = val  # data is stored internally as integer
         self.len = len_
@@ -170,7 +170,9 @@ class Bits:
         if type(k) is slice:
             return Bits(self.as_bin()[k])
         if type(k) is int:
-            if k >= self.len:
+            if k < 0:
+                k += self.len
+            if k < 0 or k >= self.len:
                 raise IndexError(k)
             return bool((1 << (self.len - k - 1)) & self.val)
         raise TypeError(type(k))
@@ -206,9 +208,13 @@ class Bits:
         return [c == '1' for c in self.as_bin()]
 
     def as_bin(self):
+        if not self.len:
+            return ''
         return f'{{0:0{self.len}b}}'.format(self.val)
 
     def as_hex(self):
+        if not self.len:
+            return ''
         # make template to pad out to number of bytes necessary to represent bits
         tmpl = f'%0{2 * (self.len // 8 + ((self.len % 8) != 0))}X'
         ret = tmpl % self.val
@@ -232,6 +238,8 @@ class Bits:
     def from_hex(cls, hex):
         if isinstance(hex, bytes):
             hex = hex.decode('ascii')
+        if hex == '':
+            return cls('')
         if not hex.startswith('0x'):
             hex = '0x' + hex
         return cls(hex)